#3236 decorator
