+++ eo-runtime/src/main/java/EOorg/EOeolang/EOsys/DispatchedNativeMethod.java
@@ -2,6 +2,6 @@ package EOorg.EOeolang.EOsys;
 
 import org.eolang.Phi;
 
-public interface DispatchedSyscall {
+public interface DispatchedNativeMethod {
     int call(Phi... params);
 }
